// src/app/features/lms/grp/components/underwriting/models/underwriting.ts
export class PolicyDocumentsDTO {
    gpd_code: number;
    gpd_rd_code: number;
    document_display: string;
    gpd_rd_sht_desc: string;
    gpd_is_doc_submitted: string;
    gpd_date_given: string;
    gpd_doc_no: number | null;
}

export class CoverTypesDTO {
    pcvt_cvt_sht_desc: string;
    pcvt_sa: number;
    pcvt_premium: number;
    pcvt_but_charge_premium: number | null;
    cvt_depend_disp: string;
    pcvt_tot_members: number;
    pcvt_load_disc: string;
    pcvt_disc_load_rate: number | null;
    pcvt_disc_load_div_fact: number | null;
    pcvt_average_period: number | null;
    pcvt_sa_per_member: number | null;
    pcvt_use_unit_rate: string;
    pcvt_unit_rate: number;
    pcvt_code: number;
    pcvt_main_rider: string;
    pcvt_cvt_code: number;
    pcvt_avg_anb: number | null;
    pcvt_fcl_amt: number | null;
    pcvt_override_fcl_amt: number | null;
    pcvt_main_sa_perc: number | null;
    pcvt_main_cover: string;
    main_cover_desc: string;
    pcvt_pct_code: number;
    pcvt_loanamt_per_mem: number | null;
    pcvt_loan_rpymt_prd: number | null;
    pcvt_loanint_per_mem: number | null;
    pcvt_savings_per_mem: number | null;
    pcvt_disc: string;
    pcvt_disc_rate: number | null;
    pcvt_disc_div_fact: number | null;
    pcvt_pmas_code: number | null;
    pcvt_pmas_sht_desc: string | null;
    pcvt_rate: number | null;
    pcvt_rate_div_fact: number;
    pcvt_use_cvr_rate: string;
    pcvt_load_age_fact: number | null;
    pcvt_override_ret_limit: number | null;
    pcvt_override_ret_lim_amt: number | null;
    pcvt_tot_member_earnings: number | null;
    pcvt_tot_orig_loan_amt: number | null;
    pcvt_mult_earnings_prd: number | null;
    pcvt_staff_desc: string | null;
    pcvt_avg_mem_earn: number | null;
    pcvt_apply_comm_exp_loading: string | null;
    pcvt_facultative: string;
    pcvt_cover_beneficiaries: string | null;
    pcvt_comm_rate: number | null;
    pcvt_comm_amt: number | null;
    pcvt_waiting_period: number | null;
    pcvt_formular: string;
    pcvt_refund_formular: string;
}

export class CoinsuranceDetailsDTO {
    pol_coin_code: number;
    pol_coin_endr_code: number;
    pol_coin_agn_code: number;
    pol_coin_pol_code: string | null;
    coinsurer_display: string | null;
    pol_coin_admin_fee: number | null;
    pol_coin_share_pct: number | null;
    pol_coinsure_leader: string | null;
    pol_coinsure_leader_share: number | null;
    follower_leader: string | null;
    follower_leader_desc: string | null;
    pol_coin_premium: number | null;
    pol_coin_srvce_fee: number | null;
    endr_coin_follower_share: number | null;
    endr_follower_prem_provided: number | null;
    endr_leader_combined: number | null;
    pol_coin_settled: string | null;
    pol_coin_comm_amt: number | null;
    btr_trans_type: string | null;
    pol_coin_balance: number | null;
    fully_settled: string | null;
    pol_coin_med_fees: number | null;
    tot_coin_sa: number | null;
    pol_coin_ov_comm: number | null;
    pol_coin_vat_amt: number | null;
    lpfp_fac_comm_rate: number | null;
    lpfp_fac_comm_amt: number | null;
}

export class CategoryDTO {
    pca_comm_rate: number | null;
    pca_comm_amt: number | null;
    pmas_sht_desc: string;
    limit_amount: number | null;
    loan_amount: number | null;
    lpag_code: number | null;
    average_cvr_earning_permember: number | null;
    maximum_type_allowed: number | null;
    minimum_amt: number | null;
    cvt_desc: string | null;
    dty_sht_desc: string | null;
    main_cover: string | null;
    cover_inbuilt: string | null;
    sum_assured_percentage: number | null;
    accelerator: number | null;
    level: number | null;
    policy_access_group: string;
    policy_category_code: number;
    short_description: string;
    category_category: string;
    policy_code: string | null;
    period: number;
    med_exempted: number | null;
    endorsement_code: string | null;
    school_code: string | null;
    premium_mask_code: number;
    premium_mask_desc: string;
    previous_category_code: number | null;
    use_cvr_rate: string;
    rate: number | null;
    category_rate_division_factor: number | null;
    average_earnings_per_member: number | null;
    average_anb: number | null;
    sum_assured_per_member: number | null;
    multiple_earnings_period: number | null;
    total_member_earnings: number | null;
    total_original_loan_amount: number | null;
    total_members: number | null;
    sum_assured: number | null;
    premium: number | null;
    base_sum_assured: number | null;
    base_premium: number | null;
    fee_amount: number | null;
    total_students: number | null;
}